Change how JS is executed on input.
We now inline everything, including scripts, and run the Javascript on the inlined document. The draw back is that the page is no executed in the contect of the rendering document, i.e. the base URL is going to change and for example ajax requests against relative URLs will fail.

The other options would have been to open the URL passed to drawURL inside a iframe directly, however we won't know if loading that URL failed as onerror doesn't trigger on a failing document.