Add font-size to root <svg> for proper rem sizing
The size of elements with properties specified in [CSS3 rem units] [1]
is determined based on the font-size of the root element.  When
appearing in an SVG document, this is the root <svg> element rather than
the <html> element which appears inside <foreignObject>.  In order to
properly size such elements, copy the font-size from the HTML document
root to the SVG document root.

1.  http://www.w3.org/TR/css3-values/#rem-unit

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>